pyproject.toml
[project]
name = "fitness-bot"
version = "0.1.0"
description = "Telegram fitness bot — calorie tracking, nutrition, workouts"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.15,<4",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "openai>=1.60,<2",
    "redis>=5.0,<6",
    "pillow>=11.0,<12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100
